restructure strapi files into a strapi-sections folder
