+++ components/strapi-sections/index.js
@@ -1,0 +1,2 @@
+export * from `./sections`
+export * from './strapi-components'
