+++ components/strapi-sections/sections.js
@@ -1,9 +1,9 @@
 import { useRouter } from "next/router"
-import RichText from "./sections/rich-text"
-import { TitleAndBody } from "./sections/title-and-body"
-import { PageTitle } from "./sections/page-title"
-import { Hero } from "./sections/hero"
-
+import {
+  TitleAndBody, 
+  PageTitle, 
+  RichText, 
+  Hero } from "./strapi-components"
 
 // Map Strapi sections to section components
 const sectionComponents = {
