+++ components/strapi-sections/strapi-components/hero/index.js
