+++ components/strapi-sections/strapi-components/index.js
@@ -1,0 +1,4 @@
+export * from './hero'
+export * from './page-title'
+export * from './rich-text'
+export * from './title-and-body'
